Make sidebar visibility configurable
