+++ euporie/core/style.py
@@ -507,18 +507,18 @@ def build_style(
         # Drop-shadow
         "drop-shadow inner": f"fg:{cp.bg.towards(cp.ansiblack, 0.3)}",
         "drop-shadow outer": f"fg:{cp.bg.towards(cp.ansiblack, 0.2)} bg:{cp.bg.towards(cp.ansiblack, 0.05)}",
-        # Sidebar
-        "sidebar": f"bg:{cp.bg.less(0.15)}",
-        "sidebar title": f"fg:{cp.hl}",
-        "sidebar title text": f"fg:default bg:{cp.bg.less(0.15).more(0.01)}",
-        "sidebar border": f"fg:{cp.bg.towards(cp.ansiblack, 0.3)}",
-        "sidebar border outer": f"bg:{cp.bg}",
-        "sidebar buttons": f"bg:{cp.bg.less(0.15)}",
-        "sidebar buttons hovered": f"fg:{cp.hl}",
-        "sidebar buttons selection": f"fg:{cp.fg} bg:{cp.hl}",
-        "sidebar buttons separator": f"fg:{cp.bg.less(0.15)} bg:{cp.bg.less(0.15)}",
-        "sidebar buttons separator selection before": f"fg:{cp.bg.less(0.15)} bg:{cp.hl}",
-        "sidebar buttons separator selection after": f"fg:{cp.hl} bg:{cp.bg.less(0.15)}",
+        # Side-bar
+        "side_bar": f"bg:{cp.bg.less(0.15)}",
+        "side_bar title": f"fg:{cp.hl}",
+        "side_bar title text": f"fg:default bg:{cp.bg.less(0.15).more(0.01)}",
+        "side_bar border": f"fg:{cp.bg.towards(cp.ansiblack, 0.3)}",
+        "side_bar border outer": f"bg:{cp.bg}",
+        "side_bar buttons": f"bg:{cp.bg.less(0.15)}",
+        "side_bar buttons hovered": f"fg:{cp.hl}",
+        "side_bar buttons selection": f"fg:{cp.fg} bg:{cp.hl}",
+        "side_bar buttons separator": f"fg:{cp.bg.less(0.15)} bg:{cp.bg.less(0.15)}",
+        "side_bar buttons separator selection before": f"fg:{cp.bg.less(0.15)} bg:{cp.hl}",
+        "side_bar buttons separator selection after": f"fg:{cp.hl} bg:{cp.bg.less(0.15)}",
         # Tabbed split
         "tabbed-split tab-bar tab inactive": f"fg:{cp.bg.more(0.1)}",
         "tabbed-split tab-bar tab inactive border": f"fg:{cp.bg.more(0.15)}",
